+++ apps/openmw/mwclass/ingredient.cpp
@@ -134,7 +134,7 @@ namespace MWClass
             if (ref->mBase->mData.mEffectID[i] < 0)
                 continue;
             MWGui::Widgets::SpellEffectParams params;
-            params.mEffectID = ref->mBase->mData.mEffectID[i];
+            params.mEffectID = static_cast<short>(ref->mBase->mData.mEffectID[i]);
             params.mAttribute = ESM::Attribute::indexToRefId(ref->mBase->mData.mAttributes[i]);
             params.mSkill = ESM::Skill::indexToRefId(ref->mBase->mData.mSkills[i]);
             params.mKnown = alchemySkill >= fWortChanceValue * (i + 1);
