+++ apps/openmw/mwgui/itemtransfer.hpp
@@ -55,7 +55,7 @@ namespace MWGui
             WorldItemModel worldItemModel(0.5f, 0.5f);
             ItemModel* const targetModel = targetView == nullptr ? &worldItemModel : targetView->getModel();
 
-            if (!targetModel->onDropItem(item.mBase, count))
+            if (!targetModel->onDropItem(item.mBase, static_cast<int>(count)))
                 return;
 
             sourceView.getModel()->moveItem(item, count, targetModel);
