+++ apps/openmw/mwgui/spellwindow.cpp
@@ -255,7 +255,7 @@ namespace MWGui
             return;
 
         mSpellView->setModel(new SpellModel(MWMechanics::getPlayer()));
-        int itemCount = mSpellView->getModel()->getItemCount();
+        int itemCount = static_cast<int>(mSpellView->getModel()->getItemCount());
         if (itemCount == 0)
             return;
