+++ apps/openmw/mwlua/uibindings.cpp
@@ -229,7 +229,10 @@ namespace MWLua
             return luaManager->uiResourceManager()->registerTexture(std::move(data));
         };
 
-        api["screenSize"] = []() { return osg::Vec2f(Settings::video().mResolutionX, Settings::video().mResolutionY); };
+        api["screenSize"] = []() {
+            return osg::Vec2f(
+                static_cast<float>(Settings::video().mResolutionX), static_cast<float>(Settings::video().mResolutionY));
+        };
 
         api["_getAllUiModes"] = [](sol::this_state thisState) {
             sol::table res(thisState, sol::create);
@@ -255,16 +258,16 @@ namespace MWLua
                           if (arg.has_value())
                               ptr = arg->ptr();
                           const std::vector<MWGui::GuiMode>& stack = windowManager->getGuiModeStack();
-                          unsigned common = 0;
+                          size_t common = 0;
                           while (common < std::min(stack.size(), newStack.size()) && stack[common] == newStack[common])
                               common++;
                           // TODO: Maybe disallow opening/closing special modes (main menu, settings, loading screen)
                           // from player scripts. Add new Lua context "menu" that can do it.
-                          for (unsigned i = stack.size() - common; i > 0; i--)
+                          for (size_t i = stack.size() - common; i > 0; i--)
                               windowManager->popGuiMode(true);
                           if (common == newStack.size() && !newStack.empty() && arg.has_value())
                               windowManager->pushGuiMode(newStack.back(), ptr);
-                          for (unsigned i = common; i < newStack.size(); ++i)
+                          for (size_t i = common; i < newStack.size(); ++i)
                               windowManager->pushGuiMode(newStack[i], ptr);
                       },
                       "Set UI modes");
